2019: Day 3 Done! MATH! YES!
i still don't get why |x+y| is not the same as |x|+|y|. i mean it's totally not the same but why is math like that. why can't it be simpler. maybe it's just too late and i should get some sleep... no can't be! I call maths a lyer and that's the reason it didn't work!